requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
Pillow>=10.0.0
gspread>=6.0.0
google-auth>=2.29.0
google-api-python-client>=2.126.0
streamlit-image-zoom>=0.0.5
streamlit-image-coordinates>=0.4.0
